requirements.txt
numpy~=1.24.3
pandas~=1.5.3
autograd~=1.6.2
tensorflow==2.11.0
tensorflow-probability==0.19.0
matplotlib~=3.7.3
scikit-survival~=0.22.0
scikit-learn~=1.3
notebook~=7.0.4
seaborn~=0.12.2
pyarrow~=13.0.0
lifelines~=0.26.4
torch~=1.13.1
torchvision~=0.14.1
pycox~=0.2.2
scipy~=1.10.1
scikit-multilearn~=0.2.0
sklearn_pandas~=2.2.0
tqdm~=4.66.2
rpy2==3.5.1
setuptools~=52.0.0
tzlocal~=5.1
statsmodels==0.14.0
bnnsurv==0.1.3
iterative-stratification==0.1.7
xgbse==0.2.3